Moves MainMenu into 'src/components'
