+++ src/components/Header.js
@@ -5,7 +5,7 @@ import {
   Visibility,
 } from 'semantic-ui-react';
 
-import { MainMenu } from '../MainMenu';
+import { MainMenu } from './MainMenu';
 import FixedMenu from '../FixedMenu';
 
 class Header extends React.Component {
